+++ resources/lang/en/prompter.php
@@ -3,6 +3,6 @@
 return [
     "bad_method_call" => "The method `:method` does not exist in this class `:class`.",
     "unsupported_input_type" => "Unsupported input type: `:type`. Supported types are: :supportedTypes",
-
+    "bad_context_method_call" => "Bad context call method: `:method` does not exist. Supported methods are: :methods",
 ];